Update README.md
I noticed that it's more productive and easier for me to write things down despite the ever-increasing integration of technology in our day-to-day lives. Instead of fearing it, I decided it would be better to utilize it to improve my productivity. 
I am building this project to convert my handwriting (from an image) to a printable document for my new portable printer.